pyproject.toml
[project]
name = "IS-GOAT-BURNING"
version = "0.1.0"
description = "This project monitors the Gävle Goat webcam feed to detect if the goat is on fire. It uses computer vision techniques to analyze the video feed and sends email and/or discord notifications if fire is detected."
authors = [
    { name = "Miłosz Białczak" }
]
dependencies = [
    "numpy==2.3.2",
    "vidgear==0.3.3",
    "yt-dlp==2025.8.11",
    "aiohttp==3.12.15",
    "pydantic-settings==2.10.1",
]
requires-python = ">=3.11"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.optional-dependencies]
cpu = [
    "opencv-python==4.11.0.86",
]
dev = [
    "ruff==0.9.2",
    "pre-commit==4.1.0",
    "pip-tools==7.5.0",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "pytest-timeout==2.3.1",
    "flake8-docstrings==1.7.0",
]
[tool.ruff]
exclude = [
    "venv",
    ".venv",
]
line-length = 128
indent-width = 4
target-version = "py311"
lint.select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "PGH",  # pygrep-hooks
    "RET",  # flake8-return
]
lint.ignore = [
    "ARG002",
]
# Enable sorting of imports
[tool.ruff.lint.isort]
force-single-line = true
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.setuptools.packages.find]
include = ["is_goat_burning*"]

[build-system]
requires = ["setuptools>=75.8.0", "wheel"]
build-backend = "setuptools.build_meta"
